Moved distro class definition to separate file
